refactor: Address type safety, API clarity, and complexity issues
This commit resolves a series of static analysis warnings and improves code quality by refactoring a high-complexity function.

- Resolves `bugprone-narrowing-conversions` by adding explicit casts to `double` in floating-point calculations to prevent data loss.

- Fixes `bugprone-implicit-widening-of-multiplication-result` by casting operands to a wider type before multiplication to prevent potential overflows.

- Improves API clarity by renaming parameters in functions flagged with `bugprone-easily-swappable-parameters` to be more descriptive.

- Reduces cognitive complexity by refactoring the `ft_glob_to_pcre` function. The logic is now broken down into smaller, well-documented helper functions, making it more readable and maintainable.

+++ check/check_parallel_hashing.c
@@ -333,7 +333,7 @@ START_TEST(test_many_files)
 	char base_path[MAX_PATH_LENGTH];
 	memset(base_path, 0, sizeof(base_path));
 	(void) snprintf(base_path, sizeof(base_path), "check/tests/many_test/base%d.dat", i);
-	create_test_file(base_path, KIBIBYTE + i * STRESS_TEST_ITERATIONS);
+	create_test_file(base_path, (size_t) KIBIBYTE + (size_t) i * STRESS_TEST_ITERATIONS);
 
 	for (int j = 1; j <= 2; j++) {
 	    char dup_path[MAX_PATH_LENGTH];
